refactor(tests): extract shared fixture_builder from golden_vault_001.rs
The 944-line generator becomes a thin caller of new
core/tests/common/fixture_builder.rs. Pure structural refactor:
no semantic changes, pinned bytes under core/tests/data/golden_vault_001/
unchanged.

Enables a sibling golden_vault_002/ fixture (next commit) to share
the same generator infrastructure without code duplication.
